% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparalelism-package.R
\docType{package}
\name{rparalelism-package}
\alias{rparalelism}
\alias{rparalelism-package}
\title{rparalelism: Sample usage of TBB vs RcppParallel}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Renato Samperio \email{renato.samperio@zurich.com}

}
\keyword{internal}
