
\name{rthAttachSharedVar}
\alias{rthAttachSharedVar}
\title{Attach a shared variable}
\description{Attach a previously created bigmemory shared variable using its descriptor file.}
\usage{rthAttachSharedVar(varName)}
\arguments{\item{varName}{Name of the shared variable.}}
\value{Nothing.}
\examples{\dontrun{rthAttachSharedVar("nThreads")}}

