
\name{rthSrcExamples}
\alias{rthSrcExamples}
\title{Source package examples}
\description{Load one of the package example scripts from the installed examples directory.}
\usage{rthSrcExamples(exName)}
\arguments{\item{exName}{Example file name.}}
\value{Nothing.}
\examples{\dontrun{rthSrcExamples("quickstart.R")}}

