% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parallel_sum.R
\name{test_parallel_sum}
\alias{test_parallel_sum}
\title{Run parallel benchmark}
\usage{
test_parallel_sum(vec = NA, times = 50, ncpus = NA, v_size = 1e+05)
}
\arguments{
\item{vec}{\code{vector} input array}

\item{times}{\code{numeric} how many times it shoud run}

\item{ncpus}{\code{numeric} number of cores to use}

\item{v_size}{\code{numeric} size of vector if not provided}
}
\description{
Run parallel benchmark
}
